Implement saving the boards to a JSON file
The feature of saving the boards has now been implemented. The current boards can be saved in a new JSON file with a name and in the directory given by the user. The files are saved in the UTF-8 encoding. Additionally, concerns with the way exceptions are handled arose. These will be addressed in one of the following commits.